Move config to own file
